[Rename]: constants.js를 constant 폴더로 이동
